# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès"
          },
          "409": {
            "description": "Email déjà utilisé"
          }
        },
        "summary": "Inscription d'un nouvel utilisateur",
        "tags": [
          "Authentification"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie"
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "summary": "Connexion d'un utilisateur",
        "tags": [
          "Authentification"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Plateforme Hôpitaux",
    "description": "API pour répertorier les hôpitaux, leurs services et permettre aux utilisateurs de les noter",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "jwt",
        "type": "http",
        "description": "Entrez votre token JWT"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Mot de passe (min 8 caractères)",
            "minLength": 8
          },
          "firstName": {
            "type": "string",
            "example": "Jean",
            "description": "Prénom de l'utilisateur"
          },
          "lastName": {
            "type": "string",
            "example": "Dupont",
            "description": "Nom de famille de l'utilisateur"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Mot de passe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}